Move test to AbstractHibernateQuerySpec
