+++ data-hibernate-jpa/src/test/groovy/io/micronaut/data/hibernate/AbstractHibernateQuerySpec.groovy
@@ -17,6 +17,7 @@ package io.micronaut.data.hibernate
 
 import io.micronaut.data.exceptions.EmptyResultException
 import io.micronaut.data.hibernate.entities.Rating
+import io.micronaut.data.hibernate.entities.RelPerson
 import io.micronaut.data.hibernate.entities.UserWithWhere
 import io.micronaut.data.jpa.repository.criteria.Specification
 import io.micronaut.data.model.Pageable
@@ -64,6 +65,10 @@ abstract class AbstractHibernateQuerySpec extends AbstractQuerySpec {
     @Shared
     @Inject
     ProductRepo productRepo
+
+    @Shared
+    @Inject
+    RelPersonRepository relPersonRepo
 
     void "test @where with nullable property values"() {
         when:
@@ -678,6 +683,50 @@ abstract class AbstractHibernateQuerySpec extends AbstractQuerySpec {
             loadedProd.longName == product.longName
         cleanup:
           productRepo.deleteAll()
+    }
+
+    void "test relations person repository and joins"() {
+        given:
+            def parent = new RelPerson()
+            parent.name = 'RelParent'
+            relPersonRepo.save(parent)
+            def child1Friend1 = new RelPerson()
+            child1Friend1.name = 'Child1Friend1'
+            relPersonRepo.save(child1Friend1)
+            def child1Friend2 = new RelPerson()
+            child1Friend2.name = 'Child1Friend2'
+            relPersonRepo.save(child1Friend2)
+            def child2Friend1 = new RelPerson()
+            child2Friend1.name = 'Child2Friend1'
+            relPersonRepo.save(child2Friend1)
+            def child1 = new RelPerson()
+            child1.name = 'Child1'
+            child1.parent = parent
+            child1.friends = [child1Friend1, child1Friend2]
+            relPersonRepo.save(child1)
+            def child2 = new RelPerson()
+            child2.name = 'Child2'
+            child2.parent = parent
+            child2.friends = [child2Friend1]
+            relPersonRepo.save(child2)
+        when:
+            def result = (List<RelPerson>) relPersonRepo.findAll(RelPersonRepository.Specifications.findRelPersonByParentAndFriends(parent.id, List.of(child1Friend1.id, child1Friend2.id)))
+        then:
+            result.size() == 1
+            result[0].id == child1.id
+        when:
+            result = (List<RelPerson>) relPersonRepo.findAll(RelPersonRepository.Specifications.findRelPersonByParentAndFriends(parent.id, List.of(child1Friend1.id, child1Friend2.id, child2Friend1.id)))
+        then:
+            result.size() == 2
+        when:
+            result = (List<RelPerson>) relPersonRepo.findAll(RelPersonRepository.Specifications.findRelPersonByChildren(List.of(child1.id, child2.id)))
+        then:
+            result.size() == 1
+            result[0].id == parent.id
+        when:
+            result = (List<RelPerson>) relPersonRepo.findAll(RelPersonRepository.Specifications.findRelPersonByChildren(List.of(child1Friend1.id, child1Friend2.id, child2Friend1.id)))
+        then:
+            result.size() == 0
     }
 
     private static Specification<Book> testJoin(String value) {
